Abandoning seperate byte list
Even though it makes copying a LineInfo or SourceInfo awkward, it seems
simpler to test for correctness if we just have the byte list in the
LineInfo structure after all and just pay the copy price during
assembly. Trying to move this out is premature optimization I think,
since its not even clear at this stage that this part of the assembly
is that much of a bottleneck (that is, I haven't profiled it yet).